@@ -2,7 +2,7 @@
 name = "jobsearch-api"
 version = "1.0.0"
 description = "REST API for Job Search Automation Platform"
-requires-python = ">=3.8"
+requires-python = ">=3.12"
 dependencies = [
     "fastapi>=0.110.0",
     "uvicorn>=0.27.0",
@@ -12,7 +12,8 @@ dependencies = [
     "python-dotenv>=1.0.0",
     "pydantic>=2.6.0",
     "alembic>=1.13.0",
-    "jobsearch @ {root:uri}/.."
+    "jobsearch @ {root:uri}/..",
+    "mcp>=0.1.0"
 ]
 
 [project.optional-dependencies]
@@ -29,7 +30,7 @@ build-backend = "hatchling.build"
 
 [tool.black]
 line-length = 88
-target-version = ["py310"]
+target-version = ["py312"]
 include = '\.pyi?$'
 exclude = '''
 (
