@@ -1,6 +1,3 @@
-"""REST API service for Job Search Automation Platform"""
-
-from .main import app
+"""Job Search API application package."""
 
 __version__ = "1.0.0"
-__all__ = ['app']
