@@ -1,53 +1,53 @@
+"""Profile-related models."""
+from typing import Optional, List
 from pydantic import BaseModel
-from typing import List, Optional
-from datetime import date
 
 class SkillBase(BaseModel):
+    """Base skill model."""
     name: str
-    proficiency: Optional[str] = "intermediate"
-    years_experience: Optional[float] = 0.0
-    categories: Optional[List[str]] = []
+    proficiency: Optional[str] = None
+    years_experience: Optional[float] = None
+    categories: Optional[List[str]] = None
 
 class SkillCreate(SkillBase):
+    """Skill creation model."""
     pass
 
 class SkillResponse(SkillBase):
+    """Skill response model."""
     id: int
     
     class Config:
         from_attributes = True
 
 class ExperienceBase(BaseModel):
+    """Base experience model."""
     company: str
     title: str
-    start_date: date
-    end_date: Optional[date] = None
-    description: str
-    highlights: List[str] = []
-    skills: List[str] = []
+    start_date: str
+    end_date: Optional[str] = None
+    description: Optional[str] = None
+    highlights: Optional[List[str]] = None
 
 class ExperienceCreate(ExperienceBase):
-    pass
+    """Experience creation model."""
+    skills: List[str] = []
 
 class ExperienceResponse(ExperienceBase):
+    """Experience response model."""
     id: int
-    skills: List[SkillResponse]
+    skills: List[SkillResponse] = []
     
     class Config:
         from_attributes = True
 
 class ProfileSummary(BaseModel):
-    full_name: Optional[str]
+    """Profile summary model."""
+    full_name: Optional[str] = None
     target_roles: List[str]
     top_skills: List[SkillResponse]
     recent_experiences: List[ExperienceResponse]
-    highlight_projects: Optional[List[str]] = []
-    career_objectives: Optional[List[str]] = []
-    
-    class Config:
-        from_attributes = True
 
 class ProfileUpdate(BaseModel):
+    """Profile update model."""
     target_roles: Optional[List[str]] = None
-    career_objectives: Optional[List[str]] = None
-    full_name: Optional[str] = None
