@@ -1,46 +1,32 @@
-"""Logging configuration and utilities."""
-
+"""Logging configuration for the jobsearch package."""
 import logging
-from pathlib import Path
 
 def setup_logging(logger_name: str) -> logging.Logger:
-    """
-    Set up logging configuration for the given module
-    
-    Args:
-        logger_name (str): Name of the logger to create
-        
-    Returns:
-        logging.Logger: Configured logger instance
-    """
-    # Create logs directory if it doesn't exist
-    log_dir = Path(__file__).parent.parent.parent / 'logs'
-    log_dir.mkdir(exist_ok=True)
-    
-    # Create logger
+    """Set up logging configuration."""
+    # Get or create logger
     logger = logging.getLogger(logger_name)
-    if not logger.handlers:
-        logger.setLevel(logging.DEBUG)
-        
-        # Create formatters
-        file_formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-        console_formatter = logging.Formatter('%(levelname)s: %(message)s')
-        
-        # File handler - logs everything to file
-        file_handler = logging.FileHandler(log_dir / f'{logger_name}.log')
-        file_handler.setLevel(logging.DEBUG)
-        file_handler.setFormatter(file_formatter)
-        
-        # Console handler - logs INFO and above to console
-        console_handler = logging.StreamHandler()
-        console_handler.setLevel(logging.INFO)
-        console_handler.setFormatter(console_formatter)
-        
-        # Add handlers to logger
-        logger.addHandler(file_handler)
-        logger.addHandler(console_handler)
     
+    # Only configure if it hasn't been configured already
+    if not logger.handlers:
+        # Set default level
+        logger.setLevel(logging.INFO)
+        
+        # Create console handler
+        handler = logging.StreamHandler()
+        handler.setLevel(logging.INFO)
+        
+        # Create formatter
+        formatter = logging.Formatter(
+            '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+        )
+        
+        # Add formatter to handler
+        handler.setFormatter(formatter)
+        
+        # Add handler to logger
+        logger.addHandler(handler)
+
     return logger
 
-# Create a default logger for the core package
-logger = setup_logging('jobsearch.core')
+# Make sure the root logger has a handler to avoid "no handler found" warnings
+logging.getLogger().addHandler(logging.NullHandler())
