@@ -1,32 +1,59 @@
-#!/usr/bin/env python3
+"""Job search strategy generation and management using core components."""
 import os
-import json
 import sys
-import random
-import time
-import argparse
+import json
+import re
 import tempfile
 from pathlib import Path
-from datetime import datetime
+from datetime import datetime, timedelta
+from typing import List, Dict, Optional, Union, Tuple
 
-from dotenv import load_dotenv
-import google.generativeai as genai
+from jobsearch.core.logging import setup_logging
+from jobsearch.core.models import (
+    Experience, Skill, TargetRole, JobCache, JobApplication
+)
+from jobsearch.core.database import get_session
+from jobsearch.core.storage import GCSManager
+from jobsearch.core.ai import AIEngine
+from jobsearch.core.markdown import MarkdownGenerator
+from jobsearch.core.monitoring import setup_monitoring
+from jobsearch.core.schemas import (
+    DailyStrategy,
+    ActionItem,
+    NetworkingTarget,
+    JobAnalysis,
+    CompanyInsight,
+    RecentActivity,
+    FocusArea,
+    WeeklyFocus,
+    ProfileData
+)
 
-# Local module imports
-from jobsearch.core.logging_utils import setup_logging
-from jobsearch.features.job_search.search import search_linkedin_jobs
-from scripts.document_generator import generate_documents_for_jobs
-from jobsearch.scripts.strategy_generator import generate_daily_strategy, generate_weekly_focus
+# Initialize core components
+logger = setup_logging('job_strategy')
+storage = GCSManager()
+ai_engine = AIEngine(feature_name='job_strategy')
+markdown = MarkdownGenerator()
+monitoring = setup_monitoring('job_strategy')
+from jobsearch.scripts.strategy_generator import (
+    generate_daily_strategy, generate_weekly_focus,
+    get_recent_applications, get_high_priority_jobs, get_profile_data
+)
 from jobsearch.features.strategy_generation.formatter import (
     format_strategy_output_markdown as format_strategy_output,
     format_strategy_output_plain
 )
-from jobsearch.features.job_search.recruiter import get_recruiter_finder
-from jobsearch.core.storage import gcs
+from jobsearch.features.job_search.recruiter import get_recruiter_finder 
+from jobsearch.features.common.storage import GCSManager
+from jobsearch.core.schemas import (
+    DailyStrategy, JobMatch, ActionItem, 
+    JobSearchResult, RecruitersByCompany,
+    RecruiterInfo
+)
 
-# Import Slack notifier
+# Try to import Slack notifier
 try:
-    from slack_notifier import get_notifier
+    from jobsearch.features.common.slack import get_notifier
     SLACK_AVAILABLE = True
 except ImportError:
     SLACK_AVAILABLE = False
@@ -42,19 +69,39 @@ def search_jobs(search_queries, job_limit=5):
     
     job_searches = []
     for query in search_queries:
-        jobs = search_linkedin_jobs(query, limit=job_limit)
-        if jobs:
-            job_searches.append({
-                "role": query,
-                "listings": jobs
-            })
+        # Create and run event loop for each query
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        try:
+            jobs = loop.run_until_complete(search_jobs_async(query, limit=job_limit))
+            if jobs:
+                job_searches.append({
+                    "role": query,
+                    "listings": jobs
+                })
+        finally:
+            loop.close()
         time.sleep(random.uniform(1, 2))  # Pause between queries
     
     logger.info(f"Found {sum(len(search['listings']) for search in job_searches)} jobs across {len(job_searches)} search queries")
     return job_searches
 
-def find_recruiters_for_jobs(job_searches, limit_per_company=2, cache_only=True):
-    """Find recruiters for companies with job listings"""
+async def search_jobs_async(query: str, location: str = None, limit: int = 5):
+    """Async helper function to search jobs"""
+    from jobsearch.features.job_search.search import search_jobs as search_jobs_core
+    return await search_jobs_core(query, location, limit=limit)
+
+def find_recruiters_for_jobs(job_searches: List[Dict], limit_per_company: int = 2, cache_only: bool = True) -> RecruitersByCompany:
+    """Find recruiters for companies with job listings
+    
+    Args:
+        job_searches: List of job search results by role
+        limit_per_company: Maximum recruiters to find per company
+        cache_only: Whether to only use cached recruiter data
+        
+    Returns:
+        RecruitersByCompany: Mapping of companies to their recruiters
+    """
     logger.info("Searching for recruiters at companies with job listings")
     
     # Get recruiter finder instance
@@ -63,8 +110,8 @@ def find_recruiters_for_jobs(job_searches, limit_per_company=2, cache_only=True)
     # Track companies we've already processed to avoid duplicates
     processed_companies = set()
     
-    # Dictionary to store recruiters by company
-    company_recruiters = {}
+    # Dictionary to store validated recruiters by company 
+    company_recruiters: Dict[str, List[RecruiterInfo]] = {}
     
     # Process each job search result
     for search in job_searches:
@@ -89,462 +136,764 @@ def find_recruiters_for_jobs(job_searches, limit_per_company=2, cache_only=True)
     logger.info(f"Found recruiters for {len(company_recruiters)} companies")
     return company_recruiters
 
-def validate_strategy_content(strategy):
-    """Validate the content of the strategy"""
-    return bool(strategy.get('daily_focus')) and bool(strategy.get('weekly_focus'))
+def validate_strategy_content(strategy: DailyStrategy) -> bool:
+    """Validate the content of the strategy
+    
+    Args:
+        strategy: DailyStrategy to validate
+        
+    Returns:
+        bool: Whether the strategy is valid
+    """
+    if not strategy:
+        return False
+        
+    try:
+        _ = DailyStrategy(**strategy.model_dump()) if isinstance(strategy, DailyStrategy) else DailyStrategy(**strategy)
+        return True
+    except Exception as e:
+        logger.warning(f"Strategy validation failed: {e}")
+        return False
 
-def enhance_with_default_content(strategy):
-    """Enhance strategy with default content"""
-    strategy['daily_focus'] = strategy.get('daily_focus', {'title': 'Default Daily Focus'})
-    strategy['weekly_focus'] = strategy.get('weekly_focus', 'Default Weekly Focus')
-    return strategy
+def create_default_strategy() -> DailyStrategy:
+    """Create a default strategy with basic content"""
+    return DailyStrategy(
+        focus_area="Apply to Priority Jobs",
+        goals=[
+            "Focus on applying to highest priority roles",
+            "Review existing applications",
+            "Research target companies"
+        ],
+        action_items=[
+            ActionItem(
+                description="Review and apply to high priority job matches",
+                priority="high",
+                deadline="EOD",
+                metrics=["3 applications submitted"]
+            ),
+            ActionItem(
+                description="Research target companies",
+                priority="medium",
+                deadline="EOD",
+                metrics=["3 companies researched"]
+            )
+        ],
+        resources_needed=[
+            "Job search platform access",
+            "Updated resume",
+            "Company research tools"
+        ],
+        success_metrics={
+            "applications": "3 quality applications submitted",
+            "research": "3 companies thoroughly researched",
+            "networking": "2 meaningful professional connections"
+        }
+    )
 
-def get_sample_jobs():
-    """Provide sample job data for fallback"""
+def enhance_with_default_content(strategy: Optional[Union[Dict, DailyStrategy]] = None) -> DailyStrategy:
+    """Enhance strategy with default content
+    
+    Args:
+        strategy: Optional existing strategy to enhance
+        
+    Returns:
+        DailyStrategy: Enhanced strategy
+    """
+    try:
+        # If we have a strategy, try to convert/validate it
+        if strategy:
+            if isinstance(strategy, DailyStrategy):
+                return strategy
+            return DailyStrategy(**strategy)
+    except Exception as e:
+        logger.warning(f"Error converting existing strategy, using default: {e}")
+    
+    # Create default strategy if conversion failed or no strategy provided
+    return create_default_strategy()
+
+def get_sample_jobs() -> List[JobMatch]:
+    """Provide sample job data for fallback
+    
+    Returns:
+        List[JobMatch]: List of sample job matches
+    """
     return [
-        {"title": "Sample Job 1", "company": "Sample Company A", "application_priority": "high"},
-        {"title": "Sample Job 2", "company": "Sample Company B", "application_priority": "medium"},
-        {"title": "Sample Job 3", "company": "Sample Company C", "application_priority": "low"}
+        JobMatch(
+            title="Senior Cloud Engineer",
+            company="Sample Tech Co A",
+            match_score=95,  # Score from 0-100
+            application_priority="high",
+            key_requirements=["AWS", "Kubernetes", "Terraform"],
+            culture_indicators=["Remote-first", "Strong engineering culture"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="DevOps Team Lead",
+            company="Sample Tech Co B",
+            match_score=90,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["CI/CD", "Team Leadership", "Cloud Platforms"],
+            culture_indicators=["Work-life balance", "Mentorship focus"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="Platform Engineer",
+            company="Sample Tech Co C",
+            match_score=75,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["Infrastructure as Code", "Cloud Architecture", "Python"],
+            culture_indicators=["Startup environment", "Innovation focused"],
+            growth_potential="medium"
+        )
     ]
 
-def get_target_roles_from_profile():
-    """Extract target job roles from the user's profile data in the database"""
-    logger.info("Retrieving target roles from profile data")
-    try:
-        # Import needed modules
-        from models import TargetRole, Experience
-        from utils import session_scope
-        from strategy_generator import get_profile_data
-        
-        # First try to get from database
-        try:
-            with session_scope() as session:
-                roles = session.query(TargetRole).all()
-                if roles:
-                    target_roles = [role.role_name for role in roles if hasattr(role, 'role_name') and role.role_name]
-                    logger.info(f"Found {len(target_roles)} target roles in database: {target_roles}")
-                    return target_roles
-        except Exception as e:
-            logger.warning(f"Error accessing database for target roles: {str(e)}")
-        
-        # If database query fails, try using profile data from strategy generator
-        profile_data = get_profile_data()
-        if profile_data and 'target_roles' in profile_data and profile_data['target_roles']:
-            target_roles = [role['title'] for role in profile_data['target_roles'] if 'title' in role and role['title']]
-            if target_roles:
-                logger.info(f"Found {len(target_roles)} target roles from profile data: {target_roles}")
-                return target_roles
-                
-        # Try to get previous job titles from user's experience (NEW FALLBACK)
-        previous_titles = []
-        
-        # First try from database
-        try:
-            with session_scope() as session:
-                experiences = session.query(Experience).order_by(Experience.end_date.desc()).all()
-                if experiences:
-                    previous_titles = [exp.title for exp in experiences if hasattr(exp, 'title') and exp.title]
-                    logger.info(f"Found {len(previous_titles)} previous job titles in database: {previous_titles}")
-        except Exception as e:
-            logger.warning(f"Error accessing database for experience: {str(e)}")
-            
-        # If database query fails, try using profile data
-        if not previous_titles and profile_data and 'experiences' in profile_data:
-            previous_titles = [exp.get('title') for exp in profile_data['experiences'] if 'title' in exp and exp['title']]
-            logger.info(f"Found {len(previous_titles)} previous job titles from profile data: {previous_titles}")
-        
-        # If we have previous titles, use them
-        if previous_titles:
-            # Limit to 3-5 most recent titles and remove duplicates while preserving order
-            seen = set()
-            unique_titles = []
-            for title in previous_titles:
-                if title not in seen:
-                    seen.add(title)
-                    unique_titles.append(title)
-            
-            return unique_titles[:5]
-        
-        # If no roles are found in profile data, analyze skills to suggest roles
-        if profile_data and 'skills' in profile_data and profile_data['skills']:
-            skill_names = [skill['skill_name'] for skill in profile_data['skills'] if 'skill_name' in skill]
-            top_skills = skill_names[:5] if len(skill_names) > 5 else skill_names
-            
-            if top_skills:
-                logger.info(f"No target roles found, attempting to analyze skills: {top_skills}")
-                target_roles = suggest_roles_from_skills(top_skills)
-                if target_roles:
-                    return target_roles
+def get_recent_applications() -> List[Dict]:
+    """Get recent job applications from database using core session.
     
+    Returns:
+        List of recent job applications with relevant data
+    """
+    try:
+        with get_session() as session:
+            # Get applications from last 30 days
+            cutoff = (datetime.now() - timedelta(days=30)).isoformat()
+            applications = session.query(JobApplication).filter(
+                JobApplication.application_date >= cutoff
+            ).order_by(JobApplication.application_date.desc()).all()
+            
+            return [app.to_dict() for app in applications]
+            
     except Exception as e:
-        logger.error(f"Error getting target roles from profile: {str(e)}")
-    
-    # Default fallback roles if nothing else works
-    fallback_roles = ["Software Engineer", "Project Manager", "Data Analyst"]
-    logger.warning(f"Using generic fallback target roles: {fallback_roles}")
-    return fallback_roles
+        logger.error(f"Error getting recent applications: {str(e)}")
+        return []
 
-def suggest_roles_from_skills(skills):
-    """Use AI to suggest job roles based on user's top skills"""
-    logger.info(f"Suggesting job roles based on skills: {skills}")
+def get_high_priority_jobs() -> List[Dict]:
+    """Get high priority jobs from database using core session.
     
+    Returns:
+        List of high priority jobs needing attention
+    """
     try:
-        skills_text = ", ".join(skills)
-        
-        prompt = f"""Based on the following professional skills, suggest 3-5 specific job titles/roles that would be a good match for a job search:
+        with get_session() as session:
+            jobs = session.query(JobCache).filter(
+                JobCache.application_priority == 'high'
+            ).order_by(
+                JobCache.match_score.desc()
+            ).limit(10).all()
+            
+            return [job.to_dict() for job in jobs]
+            
+    except Exception as e:
+        logger.error(f"Error getting high priority jobs: {str(e)}")
+        return []
 
-Skills: {skills_text}
+def get_target_roles_from_profile() -> List[str]:
+    """Get target role names from database using core session.
+    
+    Returns:
+        List of target role names to search for
+    """
+    try:
+        with get_session() as session:
+            roles = session.query(TargetRole).order_by(
+                TargetRole.priority.desc()
+            ).all()
+            
+            return [role.role_name for role in roles]
+            
+    except Exception as e:
+        logger.error(f"Error getting target roles: {str(e)}")
+        return []
 
-Provide only the job titles as a comma-separated list. Be specific and relevant to the listed skills."""
+def get_market_position(company: str) -> str:
+    """Get company's market position using core AI engine.
+    
+    Args:
+        company: Company name to analyze
         
-        model = genai.GenerativeModel('gemini-1.5-pro')
-        response = model.generate_content(
-            prompt,
-            generation_config={
-                "max_output_tokens": 100,
-                "temperature": 0.2,
-            }
+    Returns:
+        Market position description
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s market position considering:
+1. Industry standing
+2. Market share
+3. Competitive advantages
+4. Growth trajectory
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown market position"
+        
+    except Exception as e:
+        logger.error(f"Error getting market position: {str(e)}")
+        return "Unknown market position"
+
+def analyze_growth(company: str) -> str:
+    """Analyze company's growth trajectory using core AI engine.
+    
+    Args:
+        company: Company name to analyze
+        
+    Returns:
+        Growth analysis
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s growth trajectory considering:
+1. Recent expansion
+2. Hiring trends
+3. Product/service development
+4. Industry outlook
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown growth trajectory"
+        
+    except Exception as e:
+        logger.error(f"Error analyzing growth: {str(e)}")
+        return "Unknown growth trajectory"
+
+def get_development_opportunities(job: JobAnalysis) -> List[str]:
+    """Get development opportunities from job using core AI engine.
+    
+    Args:
+        job: Job analysis to extract opportunities from
+        
+    Returns:
+        List of development opportunities
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze this job's development opportunities:
+
+Title: {job.title}
+Company: {job.company}
+Requirements: {', '.join(job.key_requirements)}
+
+List 3-5 specific development opportunities in this role."""
         )
         
-        # Process the response to extract role names
-        if response and response.text:
-            # Split by common separators and clean up
-            suggested_roles = [role.strip() for role in response.text.replace('\n', ',').split(',')]
-            # Filter out empty strings
-            suggested_roles = [role for role in suggested_roles if role]
-            
-            if suggested_roles:
-                logger.info(f"AI suggested roles: {suggested_roles}")
-                return suggested_roles[:5]  # Limit to 5 roles
-    
-    except Exception as e:
-        logger.error(f"Error suggesting roles from skills: {str(e)}")
-    
-    return None
-
-def generate_and_save_strategy(job_searches, output_dir, send_slack=DEFAULT_SLACK_NOTIFICATIONS, include_recruiters=False, cache_only=True):
-    """Generate and save job search strategy"""
-    logger.info("Generating job search strategy")
-    
-    # Flatten job list
-    all_jobs = []
-    for search in job_searches:
-        all_jobs.extend(search["listings"])
-    
-    # Validate that we have enough job data to generate a meaningful strategy
-    if not all_jobs or len(all_jobs) < 3:
-        logger.warning("Insufficient job data to generate a meaningful strategy (fewer than 3 jobs)")
-        # Create fallback job data during initialization to ensure we have content
-        if not all_jobs:
-            logger.info("Using sample job data for strategy generation")
-            all_jobs = get_sample_jobs()
-    
-    # Find recruiters if requested
-    recruiters = {}
-    if include_recruiters:
-        recruiters = find_recruiters_for_jobs(job_searches, cache_only=cache_only)
-    
-    # Generate strategy
-    strategy = generate_daily_strategy(all_jobs)
-    
-    # Add recruiters and weekly focus to strategy
-    if recruiters:
-        strategy['recruiters'] = recruiters
-    strategy['weekly_focus'] = generate_weekly_focus([])
-    
-    # Validate strategy content
-    if not validate_strategy_content(strategy):
-        logger.warning("Generated strategy content is incomplete, enhancing with default content")
-        strategy = enhance_with_default_content(strategy)
-    
-    # Generate filenames with current date
-    current_date = datetime.now().strftime("%Y-%m-%d")
-    base_filename = f"strategy_{current_date}"
-    
-    # Store in GCS
-    with tempfile.NamedTemporaryFile(mode='w', delete=False) as temp_md:
-        temp_md_path = Path(temp_md.name)
-        temp_md.write(format_strategy_output(strategy, strategy['weekly_focus']))
-    
-    with tempfile.NamedTemporaryFile(mode='w', delete=False) as temp_txt:
-        temp_txt_path = Path(temp_txt.name)
-        temp_txt.write(format_strategy_output_plain(strategy, strategy['weekly_focus']))
-
-    # Upload to GCS
-    md_gcs_path = f'strategies/{base_filename}.md'
-    txt_gcs_path = f'strategies/{base_filename}.txt'
-
-    gcs.upload_file(temp_md_path, md_gcs_path)
-    gcs.upload_file(temp_txt_path, txt_gcs_path)
-
-    # Clean up temp files
-    temp_md_path.unlink()
-    temp_txt_path.unlink()
-    
-    logger.info(f"Strategy saved to GCS at {md_gcs_path} and {txt_gcs_path}")
-    
-    # Send Slack notification if enabled
-    if send_slack and SLACK_AVAILABLE and validate_strategy_content(strategy):
-        try:
-            logger.info("Sending Slack notification about generated job strategy")
-            daily_focus = strategy.get('daily_focus', {})
-            job_count = len(all_jobs)
-            high_priority_count = len([j for j in all_jobs if j.get('application_priority', '').lower() == 'high'])
-            recruiter_count = sum(len(recs) for recs in recruiters.values()) if recruiters else 0
-            
-            blocks = [
-                {
-                    "type": "header",
-                    "text": {
-                        "type": "plain_text",
-                        "text": f"🎯 Job Search Strategy for {current_date}",
-                        "emoji": True
-                    }
-                },
-                {
-                    "type": "section",
-                    "text": {
-                        "type": "mrkdwn",
-                        "text": f"*Today's Focus:* {daily_focus.get('title', 'Daily Planning')}"
-                    }
-                },
-                {
-                    "type": "section",
-                    "fields": [
-                        {
-                            "type": "mrkdwn",
-                            "text": f"*New Job Opportunities:* {job_count}"
-                        },
-                        {
-                            "type": "mrkdwn", 
-                            "text": f"*High-Priority Applications:* {high_priority_count}"
-                        }
-                    ]
-                }
+        if result:
+            # Split into list and clean up
+            opportunities = [
+                opp.strip('- ').strip()
+                for opp in result.split('\n')
+                if opp.strip()
             ]
+            return opportunities[:5]  # Limit to 5
             
-            # Add success metrics section using formatted metrics
-            if daily_focus.get('formatted_metrics'):
-                blocks.append({"type": "divider"})
-                blocks.append({
-                    "type": "header",
-                    "text": {
-                        "type": "plain_text",
-                        "text": "📊 Success Metrics",
-                        "emoji": True
-                    }
-                })
-                blocks.append({
-                    "type": "section",
-                    "text": {
-                        "type": "mrkdwn",
-                        "text": daily_focus['formatted_metrics']
-                    }
-                })
-            
-            # Add high priority job details
-            high_priority_jobs = [j for j in all_jobs if j.get('application_priority', '').lower() == 'high']
-            if high_priority_jobs:
-                blocks.append({"type": "divider"})
-                blocks.append({
-                    "type": "header",
-                    "text": {
-                        "type": "plain_text",
-                        "text": "🔥 High Priority Opportunities",
-                        "emoji": True
-                    }
-                })
-                
-                for job in high_priority_jobs[:5]:  # Limit to top 5
-                    blocks.append({
-                        "type": "section",
-                        "text": {
-                            "type": "mrkdwn",
-                            "text": (
-                                f"*<{job['url']}|{job['title']}>*\n"
-                                f"*Company:* {job['company']}\n"
-                                f"*Match Score:* {job.get('match_score', 0)}%\n"
-                                f"*Requirements:* {', '.join(job.get('key_requirements', ['None specified']))}"
-                            )
-                        }
-                    })
-            
-            # Add a link to the strategy file
-            blocks.append({"type": "divider"})
-            blocks.append({
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": f"<{md_gcs_path}|📝 View full strategy>"
-                }
-            })
-            
-            # Send the notification
-            notifier = get_notifier()
-            notifier.send_notification(
-                f"Job Search Strategy for {current_date} has been generated",
-                blocks=blocks
-            )
-            logger.info("Slack notification sent successfully")
-        except Exception as e:
-            logger.error(f"Error sending Slack notification: {str(e)}")
-    
-    return strategy, md_gcs_path, txt_gcs_path
-
-def generate_medium_article(strategy, preview_only=False):
-    """Generate a Medium article based on skills in the strategy"""
-    try:
-        logger.info("Generating Medium article based on job strategy skills")
-        
-        # Import the Medium publisher
-        from medium_publisher import MediumPublisher
-        
-        # Initialize Medium publisher
-        publisher = MediumPublisher()
-        
-        # Generate article in appropriate mode
-        if preview_only:
-            logger.info("Running Medium article generation in preview mode")
-            selected_skill = publisher.select_skill_for_article()
-            if selected_skill:
-                article_data = publisher.generate_article(selected_skill)
-                if article_data:
-                    article_path = publisher.save_article_locally(article_data)
-                    logger.info(f"Generated article preview: {article_path}")
-                    return article_path
-        else:
-            logger.info("Running Medium article generation and publication")
-            result = publisher.generate_and_publish_article()
-            logger.info(f"Article generation complete: {result}")
-            return result
-    except Exception as e:
-        logger.error(f"Error generating Medium article: {str(e)}")
-        return None
-
-def main():
-    """Main entry point for job strategy generation"""
-    logger.info("Starting job strategy generation process")
-    
-    # Add the parent directory to Python path to find local modules
-    root_dir = Path(__file__).resolve().parent.parent
-    sys.path.insert(0, str(root_dir))
-    
-    # Load environment variables
-    load_dotenv()
-    genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
-    
-    # Parse command line arguments
-    parser = argparse.ArgumentParser(description='Generate job search strategy')
-    parser.add_argument('--job-limit', type=int, default=5,
-                      help='Number of job postings to return per search query (default: 5)')
-    parser.add_argument('--search-only', action='store_true',
-                      help='Only search for jobs, do not generate strategy')
-    parser.add_argument('--strategy-only', action='store_true',
-                      help='Only generate strategy from existing job data')
-    parser.add_argument('--job-file', type=str,
-                      help='Path to JSON file with job data when using --strategy-only')
-    parser.add_argument('--no-slack', action='store_false', dest='send_slack',
-                      help='Disable Slack notifications')
-    parser.add_argument('--generate-article', action='store_true',
-                      help='Generate a Medium article based on skills in the strategy')
-    parser.add_argument('--preview-article', action='store_true',
-                      help='Generate a Medium article in preview mode (no publishing)')
-    parser.add_argument('--generate-documents', action='store_true',
-                      help='Generate documents for high-priority jobs')
-    parser.add_argument('--include-recruiters', action='store_true',
-                      help='Include recruiter search in strategy generation')
-    parser.add_argument('--cache-only', action='store_true',
-                      help='Only use cached recruiters, do not search online')
-    parser.add_argument('--no-cache-only', action='store_false', dest='cache_only',
-                      help='Allow searching for recruiters online')
-    parser.set_defaults(send_slack=DEFAULT_SLACK_NOTIFICATIONS, cache_only=True)
-    args = parser.parse_args()
-    
-    try:
-        job_searches = []
-        strategy = None
-        
-        # Determine operation mode based on arguments
-        if args.search_only and args.strategy_only:
-            logger.error("Cannot specify both --search-only and --strategy-only")
-            return 1
-        
-        # Search for jobs if not in strategy-only mode
-        if not args.strategy_only:
-            # Get target roles dynamically from user's profile instead of hardcoded values
-            search_queries = get_target_roles_from_profile()
-            
-            # Ensure we have at least 2-3 search queries for better results
-            if len(search_queries) < 2:
-                logger.warning(f"Only found {len(search_queries)} target roles, adding fallback roles")
-                additional_roles = ["Software Engineer", "Project Manager", "Data Analyst"]
-                for role in additional_roles:
-                    if role not in search_queries:
-                        search_queries.append(role)
-                        if len(search_queries) >= 3:
-                            break
-            
-            logger.info(f"Using search queries from profile data: {search_queries}")
-            job_searches = search_jobs(search_queries, args.job_limit)
-            
-            # Save job data for potential future use
-            job_data_path = os.path.join(root_dir, 'job_data.json')
-            with open(job_data_path, 'w') as f:
-                json.dump(job_searches, f, indent=2)
-            logger.info(f"Job search data saved to {job_data_path}")
-            
-            # Exit if search-only mode
-            if args.search_only:
-                logger.info("Job search completed. Exiting as requested (--search-only).")
-                return 0
-        
-        # Load job data from file if in strategy-only mode
-        elif args.strategy_only:
-            if args.job_file:
-                job_file = args.job_file
-            else:
-                job_file = os.path.join(root_dir, 'job_data.json')
-                
-            if not os.path.exists(job_file):
-                logger.error(f"Job data file not found: {job_file}")
-                return 1
-                
-            with open(job_file, 'r') as f:
-                job_searches = json.load(f)
-            logger.info(f"Loaded job data from {job_file}")
-        
-        # Generate strategy if not in search-only mode
-        if not args.search_only:
-            strategy_dir = os.path.join(root_dir, 'strategies')
-            strategy, md_path, txt_path = generate_and_save_strategy(
-                job_searches, 
-                strategy_dir,
-                args.send_slack,
-                args.include_recruiters,
-                args.cache_only
-            )
-        
-        # Generate documents if requested
-        if args.generate_documents and job_searches:
-            logger.info("Generating documents for high-priority jobs")
-            generated_docs = generate_documents_for_jobs(job_searches, filter_priority="high")
-            logger.info(f"Generated {len(generated_docs)} document sets for high-priority jobs")
-        
-        # Generate Medium article if requested
-        if (args.generate_article or args.preview_article) and strategy:
-            article_result = generate_medium_article(
-                strategy, 
-                preview_only=args.preview_article
-            )
-            if article_result:
-                logger.info(f"Medium article generation successful: {article_result}")
-            
-        logger.info("Job strategy generation process completed successfully")
-        return 0
+        return []
         
     except Exception as e:
-        logger.error(f"Failed to generate job strategy: {str(e)}", exc_info=True)
-        return 1
+        logger.error(f"Error getting development opportunities: {str(e)}")
+        return []
 
-if __name__ == "__main__":
-    sys.exit(main())
+def validate_strategy_content(strategy: DailyStrategy) -> bool:
+    """Validate the content of the strategy
+    
+    Args:
+        strategy: DailyStrategy to validate
+        
+    Returns:
+        bool: Whether the strategy is valid
+    """
+    if not strategy:
+        return False
+        
+    try:
+        _ = DailyStrategy(**strategy.model_dump()) if isinstance(strategy, DailyStrategy) else DailyStrategy(**strategy)
+        return True
+    except Exception as e:
+        logger.warning(f"Strategy validation failed: {e}")
+        return False
+
+def create_default_strategy() -> DailyStrategy:
+    """Create a default strategy with basic content"""
+    return DailyStrategy(
+        focus_area="Apply to Priority Jobs",
+        goals=[
+            "Focus on applying to highest priority roles",
+            "Review existing applications",
+            "Research target companies"
+        ],
+        action_items=[
+            ActionItem(
+                description="Review and apply to high priority job matches",
+                priority="high",
+                deadline="EOD",
+                metrics=["3 applications submitted"]
+            ),
+            ActionItem(
+                description="Research target companies",
+                priority="medium",
+                deadline="EOD",
+                metrics=["3 companies researched"]
+            )
+        ],
+        resources_needed=[
+            "Job search platform access",
+            "Updated resume",
+            "Company research tools"
+        ],
+        success_metrics={
+            "applications": "3 quality applications submitted",
+            "research": "3 companies thoroughly researched",
+            "networking": "2 meaningful professional connections"
+        }
+    )
+
+def enhance_with_default_content(strategy: Optional[Union[Dict, DailyStrategy]] = None) -> DailyStrategy:
+    """Enhance strategy with default content
+    
+    Args:
+        strategy: Optional existing strategy to enhance
+        
+    Returns:
+        DailyStrategy: Enhanced strategy
+    """
+    try:
+        # If we have a strategy, try to convert/validate it
+        if strategy:
+            if isinstance(strategy, DailyStrategy):
+                return strategy
+            return DailyStrategy(**strategy)
+    except Exception as e:
+        logger.warning(f"Error converting existing strategy, using default: {e}")
+    
+    # Create default strategy if conversion failed or no strategy provided
+    return create_default_strategy()
+
+def get_sample_jobs() -> List[JobMatch]:
+    """Provide sample job data for fallback
+    
+    Returns:
+        List[JobMatch]: List of sample job matches
+    """
+    return [
+        JobMatch(
+            title="Senior Cloud Engineer",
+            company="Sample Tech Co A",
+            match_score=95,  # Score from 0-100
+            application_priority="high",
+            key_requirements=["AWS", "Kubernetes", "Terraform"],
+            culture_indicators=["Remote-first", "Strong engineering culture"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="DevOps Team Lead",
+            company="Sample Tech Co B",
+            match_score=90,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["CI/CD", "Team Leadership", "Cloud Platforms"],
+            culture_indicators=["Work-life balance", "Mentorship focus"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="Platform Engineer",
+            company="Sample Tech Co C",
+            match_score=75,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["Infrastructure as Code", "Cloud Architecture", "Python"],
+            culture_indicators=["Startup environment", "Innovation focused"],
+            growth_potential="medium"
+        )
+    ]
+
+def get_recent_applications() -> List[Dict]:
+    """Get recent job applications from database using core session.
+    
+    Returns:
+        List of recent job applications with relevant data
+    """
+    try:
+        with get_session() as session:
+            # Get applications from last 30 days
+            cutoff = (datetime.now() - timedelta(days=30)).isoformat()
+            applications = session.query(JobApplication).filter(
+                JobApplication.application_date >= cutoff
+            ).order_by(JobApplication.application_date.desc()).all()
+            
+            return [app.to_dict() for app in applications]
+            
+    except Exception as e:
+        logger.error(f"Error getting recent applications: {str(e)}")
+        return []
+
+def get_high_priority_jobs() -> List[Dict]:
+    """Get high priority jobs from database using core session.
+    
+    Returns:
+        List of high priority jobs needing attention
+    """
+    try:
+        with get_session() as session:
+            jobs = session.query(JobCache).filter(
+                JobCache.application_priority == 'high'
+            ).order_by(
+                JobCache.match_score.desc()
+            ).limit(10).all()
+            
+            return [job.to_dict() for job in jobs]
+            
+    except Exception as e:
+        logger.error(f"Error getting high priority jobs: {str(e)}")
+        return []
+
+def get_target_roles_from_profile() -> List[str]:
+    """Get target role names from database using core session.
+    
+    Returns:
+        List of target role names to search for
+    """
+    try:
+        with get_session() as session:
+            roles = session.query(TargetRole).order_by(
+                TargetRole.priority.desc()
+            ).all()
+            
+            return [role.role_name for role in roles]
+            
+    except Exception as e:
+        logger.error(f"Error getting target roles: {str(e)}")
+        return []
+
+def get_market_position(company: str) -> str:
+    """Get company's market position using core AI engine.
+    
+    Args:
+        company: Company name to analyze
+        
+    Returns:
+        Market position description
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s market position considering:
+1. Industry standing
+2. Market share
+3. Competitive advantages
+4. Growth trajectory
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown market position"
+        
+    except Exception as e:
+        logger.error(f"Error getting market position: {str(e)}")
+        return "Unknown market position"
+
+def analyze_growth(company: str) -> str:
+    """Analyze company's growth trajectory using core AI engine.
+    
+    Args:
+        company: Company name to analyze
+        
+    Returns:
+        Growth analysis
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s growth trajectory considering:
+1. Recent expansion
+2. Hiring trends
+3. Product/service development
+4. Industry outlook
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown growth trajectory"
+        
+    except Exception as e:
+        logger.error(f"Error analyzing growth: {str(e)}")
+        return "Unknown growth trajectory"
+
+def get_development_opportunities(job: JobAnalysis) -> List[str]:
+    """Get development opportunities from job using core AI engine.
+    
+    Args:
+        job: Job analysis to extract opportunities from
+        
+    Returns:
+        List of development opportunities
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze this job's development opportunities:
+
+Title: {job.title}
+Company: {job.company}
+Requirements: {', '.join(job.key_requirements)}
+
+List 3-5 specific development opportunities in this role."""
+        )
+        
+        if result:
+            # Split into list and clean up
+            opportunities = [
+                opp.strip('- ').strip()
+                for opp in result.split('\n')
+                if opp.strip()
+            ]
+            return opportunities[:5]  # Limit to 5
+            
+        return []
+        
+    except Exception as e:
+        logger.error(f"Error getting development opportunities: {str(e)}")
+        return []
+
+def validate_strategy_content(strategy: DailyStrategy) -> bool:
+    """Validate the content of the strategy
+    
+    Args:
+        strategy: DailyStrategy to validate
+        
+    Returns:
+        bool: Whether the strategy is valid
+    """
+    if not strategy:
+        return False
+        
+    try:
+        _ = DailyStrategy(**strategy.model_dump()) if isinstance(strategy, DailyStrategy) else DailyStrategy(**strategy)
+        return True
+    except Exception as e:
+        logger.warning(f"Strategy validation failed: {e}")
+        return False
+
+def create_default_strategy() -> DailyStrategy:
+    """Create a default strategy with basic content"""
+    return DailyStrategy(
+        focus_area="Apply to Priority Jobs",
+        goals=[
+            "Focus on applying to highest priority roles",
+            "Review existing applications",
+            "Research target companies"
+        ],
+        action_items=[
+            ActionItem(
+                description="Review and apply to high priority job matches",
+                priority="high",
+                deadline="EOD",
+                metrics=["3 applications submitted"]
+            ),
+            ActionItem(
+                description="Research target companies",
+                priority="medium",
+                deadline="EOD",
+                metrics=["3 companies researched"]
+            )
+        ],
+        resources_needed=[
+            "Job search platform access",
+            "Updated resume",
+            "Company research tools"
+        ],
+        success_metrics={
+            "applications": "3 quality applications submitted",
+            "research": "3 companies thoroughly researched",
+            "networking": "2 meaningful professional connections"
+        }
+    )
+
+def enhance_with_default_content(strategy: Optional[Union[Dict, DailyStrategy]] = None) -> DailyStrategy:
+    """Enhance strategy with default content
+    
+    Args:
+        strategy: Optional existing strategy to enhance
+        
+    Returns:
+        DailyStrategy: Enhanced strategy
+    """
+    try:
+        # If we have a strategy, try to convert/validate it
+        if strategy:
+            if isinstance(strategy, DailyStrategy):
+                return strategy
+            return DailyStrategy(**strategy)
+    except Exception as e:
+        logger.warning(f"Error converting existing strategy, using default: {e}")
+    
+    # Create default strategy if conversion failed or no strategy provided
+    return create_default_strategy()
+
+def get_sample_jobs() -> List[JobMatch]:
+    """Provide sample job data for fallback
+    
+    Returns:
+        List[JobMatch]: List of sample job matches
+    """
+    return [
+        JobMatch(
+            title="Senior Cloud Engineer",
+            company="Sample Tech Co A",
+            match_score=95,  # Score from 0-100
+            application_priority="high",
+            key_requirements=["AWS", "Kubernetes", "Terraform"],
+            culture_indicators=["Remote-first", "Strong engineering culture"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="DevOps Team Lead",
+            company="Sample Tech Co B",
+            match_score=90,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["CI/CD", "Team Leadership", "Cloud Platforms"],
+            culture_indicators=["Work-life balance", "Mentorship focus"],
+            growth_potential="high"
+        ),
+        JobMatch(
+            title="Platform Engineer",
+            company="Sample Tech Co C",
+            match_score=75,  # Score from 0-100
+            application_priority="medium",
+            key_requirements=["Infrastructure as Code", "Cloud Architecture", "Python"],
+            culture_indicators=["Startup environment", "Innovation focused"],
+            growth_potential="medium"
+        )
+    ]
+
+def get_recent_applications() -> List[Dict]:
+    """Get recent job applications from database using core session.
+    
+    Returns:
+        List of recent job applications with relevant data
+    """
+    try:
+        with get_session() as session:
+            # Get applications from last 30 days
+            cutoff = (datetime.now() - timedelta(days=30)).isoformat()
+            applications = session.query(JobApplication).filter(
+                JobApplication.application_date >= cutoff
+            ).order_by(JobApplication.application_date.desc()).all()
+            
+            return [app.to_dict() for app in applications]
+            
+    except Exception as e:
+        logger.error(f"Error getting recent applications: {str(e)}")
+        return []
+
+def get_high_priority_jobs() -> List[Dict]:
+    """Get high priority jobs from database using core session.
+    
+    Returns:
+        List of high priority jobs needing attention
+    """
+    try:
+        with get_session() as session:
+            jobs = session.query(JobCache).filter(
+                JobCache.application_priority == 'high'
+            ).order_by(
+                JobCache.match_score.desc()
+            ).limit(10).all()
+            
+            return [job.to_dict() for job in jobs]
+            
+    except Exception as e:
+        logger.error(f"Error getting high priority jobs: {str(e)}")
+        return []
+
+def get_target_roles_from_profile() -> List[str]:
+    """Get target role names from database using core session.
+    
+    Returns:
+        List of target role names to search for
+    """
+    try:
+        with get_session() as session:
+            roles = session.query(TargetRole).order_by(
+                TargetRole.priority.desc()
+            ).all()
+            
+            return [role.role_name for role in roles]
+            
+    except Exception as e:
+        logger.error(f"Error getting target roles: {str(e)}")
+        return []
+
+def get_market_position(company: str) -> str:
+    """Get company's market position using core AI engine.
+    
+    Args:
+        company: Company name to analyze
+        
+    Returns:
+        Market position description
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s market position considering:
+1. Industry standing
+2. Market share
+3. Competitive advantages
+4. Growth trajectory
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown market position"
+        
+    except Exception as e:
+        logger.error(f"Error getting market position: {str(e)}")
+        return "Unknown market position"
+
+def analyze_growth(company: str) -> str:
+    """Analyze company's growth trajectory using core AI engine.
+    
+    Args:
+        company: Company name to analyze
+        
+    Returns:
+        Growth analysis
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze {company}'s growth trajectory considering:
+1. Recent expansion
+2. Hiring trends
+3. Product/service development
+4. Industry outlook
+
+Provide a one-sentence summary."""
+        )
+        return result or "Unknown growth trajectory"
+        
+    except Exception as e:
+        logger.error(f"Error analyzing growth: {str(e)}")
+        return "Unknown growth trajectory"
+
+def get_development_opportunities(job: JobAnalysis) -> List[str]:
+    """Get development opportunities from job using core AI engine.
+    
+    Args:
+        job: Job analysis to extract opportunities from
+        
+    Returns:
+        List of development opportunities
+    """
+    try:
+        result = ai_engine.generate_text(
+            prompt=f"""Analyze this job's development opportunities:
+
+Title: {job.title}
+Company: {job.company}
+Requirements: {', '.join(job.key_requirements)}
+
+List 3-5 specific development opportunities in this role."""
+        )
+        
+        if result:
+            # Split into list and clean up
+            opportunities = [
+                opp.strip('- ').strip()
+                for opp in result.split('\n')
+                if opp.strip()
+            ]
+            return opportunities[:5]  # Limit to 5
+            
+        return []
+        
+    except Exception as e:
+        logger.error(f"Error getting development opportunities: {str(e)}")
+        return []
