@@ -4,8 +4,10 @@ from pathlib import Path
 from datetime import datetime
 from jobsearch.core.logging_utils import setup_logging
 from jobsearch.core.storage import gcs
+from jobsearch.core.markdown import MarkdownGenerator
 
 logger = setup_logging('strategy_formatter')
+markdown_generator = MarkdownGenerator()
 
 def format_strategy_output_plain(strategy, weekly_focus=None):
     """Format strategy output in plain text format"""
@@ -14,6 +16,7 @@ def format_strategy_output_plain(strategy, weekly_focus=None):
     content = strategy.get('content', '')
     date = strategy.get('date', datetime.now().strftime('%Y-%m-%d'))
     jobs = strategy.get('jobs', [])
+    company_insights = strategy.get('company_insights', {})
     
     output = []
     output.append(f"JOB SEARCH STRATEGY - {date}")
@@ -24,12 +27,40 @@ def format_strategy_output_plain(strategy, weekly_focus=None):
     output.append("TODAY'S TOP JOB MATCHES")
     output.append('=' * 80)
     
-    # Add job information
+    # Add job information with company insights
     for i, job in enumerate(jobs, 1):
-        output.append(f"\nJOB {i}: {job.get('title', 'Unknown Position')} at {job.get('company', 'Unknown Company')}")
+        company = job.get('company', 'Unknown Company')
+        output.append(f"\nJOB {i}: {job.get('title', 'Unknown Position')} at {company}")
         output.append(f"Priority: {job.get('application_priority', 'Low').upper()}  |  Match Score: {job.get('match_score', 0)}%")
         output.append(f"Key Requirements: {', '.join(job.get('key_requirements', ['None specified']))}")
         output.append(f"Career Growth: {job.get('career_growth_potential', 'Unknown')}")
+        
+        # Add company insights if available
+        if company in company_insights:
+            insight = company_insights[company]
+            
+            # Add Glassdoor insights
+            if insight.get('glassdoor'):
+                glassdoor = insight['glassdoor']
+                output.append("\nCompany Culture Insights (Glassdoor):")
+                output.append(f"- Work-Life Balance: {glassdoor['work_life_balance']}")
+                output.append(f"- Management Quality: {glassdoor['management_quality']}")
+                if glassdoor['red_flags']:
+                    output.append(f"- Potential Concerns: {', '.join(glassdoor['red_flags'])}")
+                output.append(f"- Overall Assessment: {glassdoor['recommendation']}")
+            
+            # Add TechCrunch insights
+            if insight.get('techcrunch'):
+                techcrunch = insight['techcrunch']
+                output.append("\nCompany Market Insights (TechCrunch):")
+                output.append(f"- Market Position: {techcrunch['market_position']}")
+                output.append(f"- Growth Trajectory: {techcrunch['growth_trajectory']}")
+                if techcrunch['key_developments']:
+                    output.append(f"- Recent Developments: {', '.join(techcrunch['key_developments'])}")
+                output.append(f"- News Sentiment: {techcrunch['news_sentiment']}")
+                if techcrunch['recommendation']:
+                    output.append(f"- Overall Assessment: {techcrunch['recommendation']}")
+        
         output.append('-' * 50)
     
     # Add weekly focus if available
@@ -45,37 +76,17 @@ def format_strategy_output_markdown(strategy, weekly_focus=None):
     logger.info("Formatting strategy output in markdown")
     
     content = strategy.get('content', '')
-    date = strategy.get('date', datetime.now().strftime('%Y-%m-%d'))
+    date = strategy.get('date', datetime.now())
     jobs = strategy.get('jobs', [])
+    company_insights = strategy.get('company_insights', {})
     
-    output = []
-    output.append(f"# Job Search Strategy - {date}\n")
-    output.append(content)
-    output.append("\n## Today's Top Job Matches\n")
-    
-    # Add job information
-    for i, job in enumerate(jobs, 1):
-        title = job.get('title', 'Unknown Position')
-        company = job.get('company', 'Unknown Company')
-        priority = job.get('application_priority', 'Low')
-        score = job.get('match_score', 0)
-        requirements = ', '.join(job.get('key_requirements', ['None specified']))
-        growth = job.get('career_growth_potential', 'Unknown')
-        url = job.get('url', '#')
-        
-        output.append(f"### {i}. {title} at {company}\n")
-        output.append(f"**Priority**: {priority.upper()} | **Match Score**: {score}%\n")
-        output.append(f"**Key Requirements**: {requirements}\n")
-        output.append(f"**Career Growth**: {growth}\n")
-        if url != '#':
-            output.append(f"[View Job]({url})\n")
-    
-    # Add weekly focus if available
-    if weekly_focus:
-        output.append("\n## Weekly Focus\n")
-        output.append(weekly_focus)
-    
-    return '\n'.join(output)
+    return markdown_generator.generate_strategy(
+        content=content,
+        date=date if isinstance(date, datetime) else datetime.strptime(date, '%Y-%m-%d'),
+        jobs=jobs,
+        company_insights=company_insights,
+        weekly_focus=weekly_focus
+    )
 
 def store_formatted_strategy(strategy, base_filename):
     """Store formatted strategy content in GCS
