@@ -1,601 +1,214 @@
-import os
-import json
+"""Job search strategy generation using core components."""
 from pathlib import Path
 from datetime import datetime
-from models import Experience, Skill, TargetRole
-from utils import session_scope
-from logging_utils import setup_logging
-import google.generativeai as genai
-from dotenv import load_dotenv
-from gcs_utils import gcs
-import tempfile
+from typing import Dict, List, Optional
 
-# Configure Google Generative AI
-load_dotenv()
-GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')
-if not GEMINI_API_KEY:
-    raise ValueError("Please set GEMINI_API_KEY environment variable")
-genai.configure(api_key=GEMINI_API_KEY)
+from jobsearch.core.logging import setup_logging
+from jobsearch.core.database import get_session
+from jobsearch.core.storage import GCSManager
+from jobsearch.core.ai import AIEngine
+from jobsearch.core.monitoring import setup_monitoring
+from jobsearch.core.models import (
+    Experience, 
+    Skill, 
+    TargetRole, 
+    JobCache, 
+    JobApplication,
+    RecruiterContact
+)
+from jobsearch.core.markdown import MarkdownGenerator
+from jobsearch.core.schemas import (
+    DailyStrategy,
+    ActionItem,
+    FocusArea,
+    NetworkingTarget,
+    WeeklyFocus,
+    RecentActivity,
+    ProfileData,
+    CompanyInsight
+)
 
+# Initialize core components
 logger = setup_logging('strategy_generator')
+storage = GCSManager()
+ai_engine = AIEngine(feature_name='strategy_generation')
+markdown = MarkdownGenerator()
+monitoring = setup_monitoring('strategy_generation')
 
-def get_profile_data():
-    """Get profile data from the database"""
-    logger.info("Retrieving profile data from database")
+def get_recent_activity() -> RecentActivity:
+    """Get recent jobs and applications from database."""
     try:
-        with session_scope() as session:
-            experiences = session.query(Experience).order_by(Experience.end_date.desc()).all()
+        with get_session() as session:
+            # Get recent job applications 
+            recent_apps = session.query(JobApplication).join(
+                JobCache
+            ).order_by(
+                JobApplication.application_date.desc()
+            ).limit(5).all()
+            
+            app_data = []
+            for app in recent_apps:
+                app_data.append({
+                    'job_title': app.job.title,
+                    'company': app.job.company,
+                    'status': app.status,
+                    'date': app.application_date,
+                    'notes': app.notes
+                })
+                
+            # Get recent found jobs
+            recent_jobs = session.query(JobCache).order_by(
+                JobCache.first_seen_date.desc()
+            ).limit(10).all()
+            
+            job_data = []
+            for job in recent_jobs:
+                job_data.append({
+                    'title': job.title,
+                    'company': job.company,
+                    'match_score': job.match_score,
+                    'url': job.url
+                })
+                
+            return RecentActivity(
+                applications=app_data,
+                discovered_jobs=job_data
+            )
+            
+    except Exception as e:
+        logger.error(f"Error getting recent activity: {str(e)}")
+        return RecentActivity(applications=[], discovered_jobs=[])
+
+def get_profile_data() -> ProfileData:
+    """Get current profile data for strategy generation."""
+    try:
+        with get_session() as session:
+            # Get experiences in order
+            experiences = session.query(Experience).order_by(
+                Experience.end_date.desc(),
+                Experience.start_date.desc()
+            ).all()
+            
+            experience_data = []
+            for exp in experiences:
+                experience_data.append({
+                    'company': exp.company,
+                    'title': exp.title,
+                    'description': exp.description,
+                    'skills': [skill.skill_name for skill in exp.skills]
+                })
+            
+            # Get unique skills
             skills = session.query(Skill).all()
-            roles = session.query(TargetRole).all()
+            skill_names = [skill.skill_name for skill in skills]
             
-            profile_data = {
-                'experiences': [
-                    {
-                        'title': exp.title,
-                        'company': exp.company,
-                        'start_date': exp.start_date,
-                        'end_date': exp.end_date,
-                        'description': exp.description
-                    } for exp in experiences
-                ],
-                'skills': [
-                    {
-                        'skill_name': getattr(skill, 'skill_name', 'Unknown'),
-                        'years': getattr(skill, 'years', 0),
-                        'proficiency': getattr(skill, 'proficiency', 'Beginner')
-                    } for skill in skills
-                ],
-                'target_roles': [
-                    {
-                        'title': getattr(role, 'role_name', 'Unknown'),
-                        'industry': getattr(role, 'industry', 'Tech'),
-                        'salary_range': getattr(role, 'salary_range', ''),
-                        'priority': getattr(role, 'priority', 1)
-                    } for role in roles
-                ]
-            }
+            # Get target roles
+            roles = session.query(TargetRole).order_by(
+                TargetRole.priority.desc()
+            ).all()
+            target_roles = [role.role_name for role in roles]
             
-            logger.info("Successfully retrieved profile data")
-            return profile_data
+            return ProfileData(
+                experiences=experience_data,
+                skills=skill_names,
+                target_roles=target_roles
+            )
+            
     except Exception as e:
-        logger.error(f"Error retrieving profile data: {str(e)}")
-        return {'experiences': [], 'skills': [], 'target_roles': []}
+        logger.error(f"Error getting profile data: {str(e)}")
+        return ProfileData(experiences=[], skills=[], target_roles=[])
 
-def store_strategy(strategy, base_filename):
-    """Store strategy content in GCS"""
+def get_networking_targets() -> List[NetworkingTarget]:
+    """Get potential networking targets from database."""
     try:
-        # Generate both markdown and plain text versions
-        from strategy_formatter import (
-            format_strategy_output_markdown,
-            format_strategy_output_plain
-        )
-
-        markdown_content = format_strategy_output_markdown(strategy, strategy.get('weekly_focus'))
-        plain_content = format_strategy_output_plain(strategy, strategy.get('weekly_focus'))
-
-        # Store in GCS
-        with tempfile.NamedTemporaryFile(mode='w', delete=False) as temp_md:
-            temp_md_path = Path(temp_md.name)
-            temp_md.write(markdown_content)
-        
-        with tempfile.NamedTemporaryFile(mode='w', delete=False) as temp_txt:
-            temp_txt_path = Path(temp_txt.name)
-            temp_txt.write(plain_content)
-
-        # Upload to GCS
-        md_gcs_path = f'strategies/{base_filename}.md'
-        txt_gcs_path = f'strategies/{base_filename}.txt'
-
-        gcs.upload_file(temp_md_path, md_gcs_path)
-        gcs.upload_file(temp_txt_path, txt_gcs_path)
-
-        # Clean up temp files
-        temp_md_path.unlink()
-        temp_txt_path.unlink()
-
-        return md_gcs_path, txt_gcs_path
+        with get_session() as session:
+            # Get identified but not contacted recruiters
+            recruiters = session.query(RecruiterContact).filter(
+                RecruiterContact.status == 'identified'
+            ).all()
+            
+            targets = []
+            for rec in recruiters:
+                targets.append(NetworkingTarget(
+                    name=rec.name,
+                    title=rec.title,
+                    company=rec.company,
+                    source=rec.source,
+                    url=rec.url,
+                    notes=rec.notes
+                ))
+                
+            return targets
+            
     except Exception as e:
-        logger.error(f"Error storing strategy: {str(e)}")
-        raise
+        logger.error(f"Error getting networking targets: {str(e)}")
+        return []
 
-def read_strategy(date):
-    """Read strategy content from GCS for a given date
-    
-    Args:
-        date (str): Date in YYYY-MM-DD format
-    Returns:
-        dict: Strategy content with 'markdown' and 'text' keys, or None if not found
-    """
-    from gcs_utils import gcs
-    
-    md_content, txt_content = gcs.read_strategy(date)
-    if md_content is None:
+async def generate_daily_strategy() -> DailyStrategy:
+    """Generate daily job search strategy using core AI engine."""
+    try:
+        logger.info("Generating daily job search strategy")
+        monitoring.increment('strategy_generation')
+        
+        # Get input data
+        recent = get_recent_activity()
+        profile = get_profile_data()
+        targets = get_networking_targets()
+        
+        # Use AI engine for strategy generation
+        strategy = await ai_engine.generate(
+            prompt=f"""Generate a daily job search strategy based on:
+
+RECENT ACTIVITY
+Applications: {len(recent.applications)} in last week
+New Jobs Found: {len(recent.discovered_jobs)} matching profile
+Latest Application: {recent.applications[0].get('status') if recent.applications else 'None'}
+
+CURRENT PROFILE
+Skills: {', '.join(profile.skills[:5])}... ({len(profile.skills)} total)
+Target Roles: {', '.join(profile.target_roles)}
+Latest Experience: {profile.experiences[0].get('title')} at {profile.experiences[0].get('company')}
+
+NETWORKING
+Potential Contacts: {len(targets)} identified recruiters/contacts""",
+            output_type=DailyStrategy
+        )
+        
+        if strategy:
+            monitoring.track_success('strategy_generation')
+            # Save strategy to markdown file
+            strategy_path = f"strategies/strategy_{datetime.now().strftime('%Y-%m-%d')}.md"
+            strategy_md = markdown.format_strategy(strategy)
+            storage.write_file(strategy_path, strategy_md)
+            logger.info(f"Strategy saved to {strategy_path}")
+            return strategy
+            
+        monitoring.track_failure('strategy_generation')
+        logger.error("Failed to generate strategy")
         return None
         
-    return {
-        'markdown': md_content,
-        'text': txt_content
-    }
-
-def format_success_metrics(metrics):
-    """Format success metrics text properly"""
-    formatted_text = ""
-    for metric in metrics:
-        # Remove any double asterisks
-        metric = metric.replace("**", "*")
-        # Ensure proper formatting for section headers
-        if metric.startswith("*") and not metric.endswith("*"):
-            metric = metric.replace("*", "", 1)  # Remove leading asterisk
-        formatted_text += f"• {metric}\n"
-    return formatted_text.strip()
-
-def generate_daily_strategy(sorted_jobs):
-    """Generate a daily strategy based on the analyzed jobs"""
-    logger.info("Generating daily strategy")
-    
-    # Get profile data
-    profile_data = get_profile_data()
-    
-    # Check if we have enough job data
-    if not sorted_jobs or len(sorted_jobs) < 2:
-        logger.warning(f"Limited job data available: {len(sorted_jobs) if sorted_jobs else 0} jobs")
-        return create_fallback_strategy(profile_data)
-    
-    # Construct the prompt for Gemini
-    job_list = ""
-    for i, job in enumerate(sorted_jobs[:10], 1):  # Limit to top 10 jobs
-        job_list += f"Job {i}: {job['title']} at {job['company']}\n"
-        job_list += f"   Match Score: {job.get('match_score', 0)}\n"  # Use get() with default value
-        job_list += f"   Application Priority: {job.get('application_priority', 'Medium')}\n"  # Use get() with default
-        job_list += f"   Key Requirements: {', '.join(job.get('key_requirements', ['Not specified']))}\n"  # Use get() with default
-        job_list += f"   Career Growth Potential: {job.get('career_growth_potential', 'Unknown')}\n\n"  # Use get() with default
-    
-    # Fix skill_name attribute reference
-    skills_list = ", ".join([f"{skill.get('skill_name', skill.get('name', 'Unknown'))} ({skill.get('proficiency', 'N/A')})" for skill in profile_data['skills']])
-    target_roles_list = ", ".join([role['title'] for role in profile_data['target_roles']]) if profile_data['target_roles'] else "Cloud/DevOps Roles"
-    
-    prompt = f"""You are a career strategist helping to create a daily job search strategy. Based on today's job matches, create a compelling strategy document with the following sections:
-
-1. Today's Focus: A concise strategy for today's job search approach
-2. High-Priority Applications: List 1-3 jobs that should be applied to today with a brief reason for each
-3. Skill Alignment: Identify skills from the candidate's profile that align well with today's job matches
-4. Areas for Improvement: Suggest 1-2 skill areas to develop based on gaps in job requirements
-5. Weekly Focus Update: Suggest adjustments to the weekly focus based on today's job trends
-
-Candidate Skills: {skills_list}
-Target Roles: {target_roles_list}
-
-Today's Top Job Matches:
-{job_list}
-
-Today's date: {datetime.now().strftime('%Y-%m-%d')}
-
-Format your response as detailed paragraphs for each section. Be specific and actionable."""
-
-    try:
-        model = genai.GenerativeModel('gemini-1.5-pro')
-        response = model.generate_content(
-            prompt,
-            generation_config={
-                "max_output_tokens": 2000,
-                "temperature": 0.7,
-            }
-        )
-        
-        # Parse the content to extract daily focus
-        content = response.text
-        daily_focus = extract_daily_focus(content)
-        
-        strategy = {
-            'date': datetime.now().strftime('%Y-%m-%d'),
-            'content': content,
-            'jobs': sorted_jobs[:10],  # Include top 10 jobs in the strategy
-            'daily_focus': daily_focus
-        }
-        
-        if not validate_strategy_content(strategy):
-            logger.warning("Generated strategy content is incomplete, falling back to defaults")
-            strategy = {
-                'daily_focus': {'title': 'Daily Planning'},
-                'success_metrics': [],
-                'content': ''
-            }
-        
-        # Format success metrics properly
-        if 'success_metrics' in strategy['daily_focus']:
-            strategy['daily_focus']['success_metrics'] = [
-                metric.strip() for metric in strategy['daily_focus']['success_metrics']
-                if metric and metric.strip()
-            ]
-            formatted_metrics = format_success_metrics(strategy['daily_focus']['success_metrics'])
-            strategy['daily_focus']['formatted_metrics'] = formatted_metrics
-
-        logger.info("Successfully generated daily strategy")
-        return strategy
     except Exception as e:
-        logger.error(f"Error generating daily strategy: {str(e)}")
-        return create_fallback_strategy(profile_data, sorted_jobs[:10] if sorted_jobs else [])
-
-def extract_daily_focus(content):
-    """Extract daily focus information from strategy content"""
-    focus = {'title': 'Daily Planning'}
-    
-    # Try to extract the title
-    if "Today's Focus:" in content:
-        lines = content.split('\n')
-        for i, line in enumerate(lines):
-            if "Today's Focus:" in line:
-                # Try to get the title from this line
-                title = line.split("Today's Focus:", 1)[1].strip()
-                if not title and i + 1 < len(lines):
-                    # If empty, try the next line
-                    title = lines[i + 1].strip()
-                if title:
-                    focus['title'] = title
-                break
-    
-    # Try to extract success metrics
-    success_metrics = []
-    if "Success Metrics:" in content or "High-Priority Applications:" in content:
-        lines = content.split('\n')
-        capture = False
-        for line in lines:
-            if "Success Metrics:" in line or "High-Priority Applications:" in line:
-                capture = True
-                continue
-            if capture and line.strip() and line.strip()[0] in ['-', '•', '*'] and len(line.strip()) > 2:
-                metric = line.strip()[1:].strip()
-                if metric:
-                    success_metrics.append(metric)
-            # Stop capturing if we hit another section header
-            if capture and line.strip() and line.strip().endswith(':') and not line.strip().startswith('-'):
-                break
-        
-    if success_metrics:
-        focus['success_metrics'] = success_metrics
-        
-    return focus
-
-def create_fallback_strategy(profile_data, jobs=None):
-    """Create a fallback strategy when there's insufficient job data"""
-    logger.info("Creating fallback strategy due to insufficient job data")
-    
-    # Default focus based on profile data
-    target_roles = [role['title'] for role in profile_data.get('target_roles', [])]
-    if not target_roles:
-        target_roles = ["Cloud Architect", "DevOps Engineer", "Cloud Engineer"]
-    
-    top_skills = [skill['skill_name'] for skill in profile_data.get('skills', [])[:5]]
-    if not top_skills:
-        top_skills = ["AWS", "Terraform", "Kubernetes", "Python", "CI/CD"]
-    
-    # Craft a meaningful strategy with networking and skill development focus
-    current_date = datetime.now().strftime('%Y-%m-%d')
-    
-    # Create a default content with real strategy value
-    content = f"""## Today's Focus: Network and Skill Development
-*Why*: Without specific job opportunities identified, focus on expanding your professional network and enhancing your technical skills to position yourself for future opportunities.
-
-### Success Metrics
-- [ ] Connect with 3 professionals in your target industry
-- [ ] Complete 1 hour of focused learning on a key skill
-- [ ] Update your professional profiles with recent achievements
-- [ ] Research 3 companies of interest
-
-## Morning Tasks
-
-### High Priority
-1. **LinkedIn Networking** ⏱️ 60min  
-   *Why*: Building relationships with professionals in your field increases visibility and uncovers hidden opportunities.
-
-### Medium Priority
-1. **Skill Development** ⏱️ 90min  
-   *Why*: Continuous learning keeps you competitive in the rapidly evolving tech landscape.
-
-## Afternoon Tasks
-
-### High Priority
-1. **Research Target Companies** ⏱️ 60min  
-   *Why*: Understanding company culture and needs helps tailor future applications.
-
-### Medium Priority
-1. **Update Online Profiles** ⏱️ 45min  
-   *Why*: Ensures your professional presence reflects your most recent skills and achievements.
-
-## Target Roles & Current Opportunities
-
-### {target_roles[0] if target_roles else "Cloud Architect"}
-*Why*: Aligns with your core technical expertise and offers strong career growth potential.
-
-#### Key Skills to Emphasize
-- {', '.join(top_skills[:3] if top_skills else ["Terraform", "AWS", "Kubernetes"])}
-- Architecture Design
-- Technical Leadership
-
-#### Target Companies
-- AWS
-- Google Cloud
-- HashiCorp
-- Fast-growing tech startups
-
-## Networking Strategy
-**Daily Connection Target**: 3
-
-### Platforms
-- LinkedIn
-- GitHub
-- Industry Slack communities
-
-### Outreach Template
-```
-Hi [Name],
-
-I came across your profile and was impressed by your work in [specific area]. I'm currently a [Your Role] specializing in [your specialty] and I'm interested in connecting with other professionals in the [target industry] space.
-
-Best regards,
-[Your Name]
-```
-
-## Skill Development Plan
-
-### Current Focus: {top_skills[0] if top_skills else "Cloud Architecture"}
-- **Goal**: Deepen knowledge of advanced features and best practices
-- **Timeline**: Ongoing
-- **Status**: In Progress
-
-### Current Focus: {top_skills[1] if len(top_skills) > 1 else "CI/CD"}
-- **Goal**: Explore automation capabilities and integration patterns
-- **Timeline**: 2 weeks
-- **Status**: Planned
-
-## Application Strategy
-**Daily Target**: Quality over quantity
-
-### Quality Checklist
-- [ ] Tailored resume and cover letter
-- [ ] Quantifiable achievements highlighted
-- [ ] Keywords optimized for ATS
-- [ ] Company research completed
-"""
-
-    # Create the strategy object
-    strategy = {
-        'date': current_date,
-        'content': content,
-        'jobs': jobs if jobs else [],
-        'daily_focus': {
-            'title': 'Network and Skill Development',
-            'success_metrics': [
-                "Connect with 3 professionals in your target industry",
-                "Complete 1 hour of focused learning on a key skill",
-                "Update your professional profiles with recent achievements",
-                "Research 3 companies of interest"
-            ]
-        }
-    }
-    
-    return strategy
-
-def generate_weekly_focus(strategies_of_week):
-    """Generate a weekly focus based on the daily strategies of the past week"""
-    logger.info("Generating weekly focus")
-    
-    if not strategies_of_week:
-        logger.warning("No daily strategies provided for weekly focus generation")
-        return "No data available to generate weekly focus."
-    
-    # Compile job data from all strategies
-    all_jobs = []
-    for strategy in strategies_of_week:
-        all_jobs.extend(strategy.get('jobs', []))
-    
-    # Count job titles and companies
-    job_titles = {}
-    companies = {}
-    skills_required = {}
-    
-    for job in all_jobs:
-        # Count job titles
-        title = job.get('title', '').lower()
-        job_titles[title] = job_titles.get(title, 0) + 1
-        
-        # Count companies
-        company = job.get('company', '').lower()
-        companies[company] = companies.get(company, 0) + 1
-        
-        # Count required skills
-        for req in job.get('key_requirements', []):
-            req_lower = req.lower()
-            skills_required[req_lower] = skills_required.get(req_lower, 0) + 1
-    
-    # Sort by frequency
-    top_titles = sorted(job_titles.items(), key=lambda x: x[1], reverse=True)[:5]
-    top_companies = sorted(companies.items(), key=lambda x: x[1], reverse=True)[:5]
-    top_skills = sorted(skills_required.items(), key=lambda x: x[1], reverse=True)[:10]
-    
-    # Create prompt for Gemini
-    titles_text = "\n".join([f"- {title}: {count} occurrences" for title, count in top_titles])
-    companies_text = "\n".join([f"- {company}: {count} occurrences" for company, count in top_companies])
-    skills_text = "\n".join([f"- {skill}: {count} occurrences" for skill, count in top_skills])
-    
-    prompt = f"""As a career strategist, analyze the past week's job market data and create a weekly focus strategy with these sections:
-
-1. Market Trends: Analyze patterns in job titles and companies
-2. In-Demand Skills: Identify the most sought-after skills this week
-3. Weekly Focus: Recommend a focused approach for next week's job search
-4. Skill Development: Suggest 2-3 skills to prioritize developing
-5. Application Strategy: Recommend which types of roles to prioritize
-
-Past Week's Data:
-
-Most Common Job Titles:
-{titles_text}
-
-Most Common Companies:
-{companies_text}
-
-Most Requested Skills:
-{skills_text}
-
-Format your response as detailed paragraphs for each section. Be specific and actionable."""
-
-    try:
-        model = genai.GenerativeModel('gemini-1.5-pro')
-        response = model.generate_content(
-            prompt,
-            generation_config={
-                "max_output_tokens": 2000,
-                "temperature": 0.7,
-            }
-        )
-        
-        logger.info("Successfully generated weekly focus")
-        return response.text
-    except Exception as e:
-        logger.error(f"Error generating weekly focus: {str(e)}")
-        return "Unable to generate weekly focus due to an error."
-
-def generate_strategy(jobs=None, include_weekly_focus=True):
-    """Generate a complete job search strategy"""
-    try:
-        # Get profile data
-        profile_data = get_profile_data()
-
-        # Generate strategy content
-        strategy = generate_daily_strategy(jobs) if jobs else create_fallback_strategy(profile_data)
-
-        # Add weekly focus if requested
-        if include_weekly_focus:
-            weekly_focus = generate_weekly_focus([strategy])  # Using current strategy for context
-            strategy['weekly_focus'] = weekly_focus
-
-        # Store strategy in GCS
-        current_date = datetime.now().strftime('%Y-%m-%d')
-        base_filename = f'strategy_{current_date}'
-        md_path, txt_path = store_strategy(strategy, base_filename)
-
-        logger.info(f"Successfully generated and stored strategy in GCS")
-        return strategy, md_path, txt_path
-
-    except Exception as e:
+        monitoring.track_error('strategy_generation', str(e))
         logger.error(f"Error generating strategy: {str(e)}")
-        raise
+        return None
 
-def format_strategy_output(strategy, include_weekly_focus=True):
-    """Format strategy output in markdown format"""
-    logger.info("Formatting strategy output in markdown")
-    
-    content = strategy.get('content', '')
-    date = strategy.get('date', datetime.now().strftime('%Y-%m-%d'))
-    jobs = strategy.get('jobs', [])
-    weekly_focus = strategy.get('weekly_focus', '')
-    
-    output = []
-    output.append(f"# Job Search Strategy - {date}\n")
-    output.append(content)
-    output.append("\n## Today's Top Job Matches\n")
-    
-    # Add job information
-    for i, job in enumerate(jobs, 1):
-        title = job.get('title', 'Unknown Position')
-        company = job.get('company', 'Unknown Company')
-        priority = job.get('application_priority', 'Low')
-        score = job.get('match_score', 0)
-        requirements = ', '.join(job.get('key_requirements', ['None specified']))
-        growth = job.get('career_growth_potential', 'Unknown')
-        url = job.get('url', '#')
-        
-        output.append(f"### {i}. {title} at {company}\n")
-        output.append(f"**Priority**: {priority.upper()} | **Match Score**: {score}%\n")
-        output.append(f"**Key Requirements**: {requirements}\n")
-        output.append(f"**Career Growth**: {growth}\n")
-        if url != '#':
-            output.append(f"[View Job]({url})\n")
-    
-    # Add weekly focus if available
-    if include_weekly_focus and weekly_focus:
-        output.append("\n## Weekly Focus\n")
-        output.append(weekly_focus)
-    
-    # Add recruiters if available
-    if 'recruiters' in strategy and strategy['recruiters']:
-        output.append("\n## Company Recruiters\n")
-        output.append("Connect with these recruiters to expand your network and get insider information on job openings.\n")
-        
-        for company, recruiters in strategy['recruiters'].items():
-            output.append(f"### {company}\n")
-            for recruiter in recruiters:
-                name = recruiter.get('name', 'Unknown')
-                title = recruiter.get('title', 'Unknown')
-                url = recruiter.get('url', '')
-                
-                if url:
-                    output.append(f"- [{name}]({url}) - {title}")
-                else:
-                    output.append(f"- **{name}** - {title}")
-                    
-                if 'status' in recruiter:
-                    output.append(f"  - Status: {recruiter['status']}")
-            output.append("")
-    
-    return '\n'.join(output)
-
-def format_strategy_output_plain(strategy, include_weekly_focus=True):
-    """Format strategy output in plain text format"""
-    logger.info("Formatting strategy output in plain text")
-    
-    content = strategy.get('content', '')
-    date = strategy.get('date', datetime.now().strftime('%Y-%m-%d'))
-    jobs = strategy.get('jobs', [])
-    weekly_focus = strategy.get('weekly_focus', '')
-    
-    output = []
-    output.append(f"JOB SEARCH STRATEGY - {date}")
-    output.append('=' * 80)
-    output.append('')
-    output.append(content)
-    output.append('')
-    output.append("TODAY'S TOP JOB MATCHES")
-    output.append('=' * 80)
-    
-    # Add job information
-    for i, job in enumerate(jobs, 1):
-        output.append(f"\nJOB {i}: {job.get('title', 'Unknown Position')} at {job.get('company', 'Unknown Company')}")
-        output.append(f"Priority: {job.get('application_priority', 'Low').upper()}  |  Match Score: {job.get('match_score', 0)}%")
-        output.append(f"Key Requirements: {', '.join(job.get('key_requirements', ['None specified']))}")
-        output.append(f"Career Growth: {job.get('career_growth_potential', 'Unknown')}")
-        output.append('-' * 50)
-    
-    # Add weekly focus if available
-    if include_weekly_focus and weekly_focus:
-        output.append("\nWEEKLY FOCUS")
-        output.append('=' * 80)
-        output.append(weekly_focus)
-    
-    # Add recruiters if available
-    if 'recruiters' in strategy and strategy['recruiters']:
-        output.append("\nRECRUITERS FOUND")
-        output.append('=' * 80)
-        for company, recruiters in strategy['recruiters'].items():
-            output.append(f"\n{company}:")
-            for recruiter in recruiters:
-                output.append(f"- {recruiter.get('name', 'Unknown')}, {recruiter.get('title', 'Unknown')}")
-                if 'url' in recruiter:
-                    output.append(f"  URL: {recruiter['url']}")
-    
-    return '\n'.join(output)
-
-def main():
-    """Main entry point"""
+async def main() -> int:
+    """Main entry point."""
     try:
-        strategy, md_path, txt_path = generate_strategy()
-        print(f"Strategy generated successfully:")
-        print(f"Markdown version: {md_path}")
-        print(f"Text version: {txt_path}")
-        return 0
+        logger.info("Starting strategy generation")
+        strategy = await generate_daily_strategy()
+        
+        if strategy:
+            logger.info("Successfully generated daily strategy")
+            return 0
+            
+        logger.error("Failed to generate strategy")
+        return 1
+        
     except Exception as e:
-        print(f"Error: {str(e)}")
+        logger.error(f"Error in strategy generation: {str(e)}", exc_info=True)
         return 1
 
 if __name__ == "__main__":
-    exit(main())
+    import asyncio
+    exit(asyncio.run(main()))
